⏪️ Revert the introduction of buffer-local keymaps
Using the ft field in Lazy keys messes up the plugin
lazy-loading. When a plugin is lazy-loaded due to a Lazy key,
only this key is available for the plugin on the current buffer
until it is reloaded, the other plugin Lazy keys are not.

Besides, some plugin keymaps got broken when introducing general
buffer-local keymaps (e.g. "_" in Oil).

Hence I'm giving up buffer-local keymaps as it causes more
trouble than anything.