⚡️ Make some Telescope keymaps start in normal mode
At the moment, all Telescope keymaps start in insert mode. However,
this doesn't make sens for some of them, so I'm trying to change that
with the policy of keeping the find-keymaps in insert mode by default,
and the others in normal mode as Telescope is more used like a
selection menu in that case.