🐛 Fix Oil snippet completion triggers
With the latest fix of Oil snippets, after writting a file-type
snippet, when typing a second one right away the completion would
not trigger. This makes the purpose of having a line return in the
snippet useless. The alternative of typing <Esc> and <o> gives the
same output but the completion works right away. Hence, I removed
the line return from the snippet and the trailing insert node in
Oil buffer snippets.