🩹 Re-use <Tab> for Copilot suggestions
I don't like the custom implementation of the "accept whole suggestion"
feature in Copilot (create a flicker on the screen) compared to the one
automatically created, so this re-introduces the latter. Besides, this
commit makes "accept-word" mapped to <C-CR>, which works very nicely
with <C-BS> as "delete word". New keymaps are similar to the new ones
used in zsh-autosuggestions.