♻️ Replace ftplugin colorcolumn management by editorconfig
This solution is a lot (lot) simpler than the ftplugin one, so it's more
scalable & easier to maintain. Besides, it integrates directly with
Neovim features (with the editorconfig support). Last but not least, it
also adds the support of the textwidth feature, which I didn't use
previously but which is nice!