🔥 Remove Harpoon plugin (again)
I've been changing my mind a lot about this one, but I'm again removing
it, for one main reason: in practice, I really don't use it. There are
some features that might be cool to use (e.g. pin a virtual environement
directory to easily search in it, even if it's git-ignored), but this
can also be done by cd-ing into it and opening a new Neovim instance. On
the other hand, Harpoon brings a notion of state in Neovim that I don't
really like and which I don't maintain so I can't use it properly.