♻️ Replace conf.nvim by local module
I've encountered some trouble when trying to use conf.nvim in the
plugin specification of nvim-web-devicons (module not found error). I
don't know how to deal with it (or even if it's possible to do with
lazy.nvim), so the solution I found is to re-implement the conf.nvim
plugin in a local plugin, as it was done originally. Since it is a kind
of "meta" plugin (which doesn't do stuff as lazy.nvim expects it to and
needs to be "setup" during the other plugins specifications), it might
be a bit more coherent/natural that way, with a local module not handled
by lazy.nvim, just like the theme module.