⏪️ Re-introduce Telescope's find_lines picker (revert 53b78674)
In the end, I find Snacks's pickers as opinionated as Telescope's (e.g.
to use keymaps like <C-u> to delete a line, or <C-f>, <C-b> to move
character), so it is also a bit painful to configure, so I prefer to
stick to Telescope, as it's already configured to my taste.